requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
supabase==2.7.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & settings
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9

# HTTP client (used by test script)
requests==2.32.3